Make nests explode when near
